Allow the browser to cache data on node-js
This patch adds the option to to use the `Etag` and `If-None-Match`
headers with the node.js server installation. This feature can be
enabled via the "use-browser-cache" option. If this option is set to
`yes` all `GET` requests will have an `Etag` header that contians the
MD5 hash of the transmitted data and the headers.

If the browser sends an 'If-None-Match' header with the next request the
node.js server will check if the hash matches the current content. If
that is the case, only a 304 header will be sent to instruct the browser
to use the already cache information.

This reduces the bandwidth usage and startup time of TiddlyWiki. In
combination with the external JavaScript template the transmitted data
can be reduced significantly.

There are the values of a my little test wiki using this with the
root-tiddler set to `$:/core/save/all-external-js`:

|                              |   gzip=no   |   zip=yes   |
|------------------------------|-------------|-------------|
| First page load (no cache)   | 3,04 MiByte | 2,43 MiByte |
| Next page load (using cache) |  955 kiByte |  329 kiByte |